Clean project details file
