+++ package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "@heroicons/react": "^2.0.16",
     "@next-auth/prisma-adapter": "^1.0.5",
-    "@prisma/client": "^4.12.0-integration-rtld-deepbind.3",
+    "@prisma/client": "^4.11.0",
     "@radix-ui/react-avatar": "^1.0.2",
     "@radix-ui/react-dialog": "^1.0.3",
     "@radix-ui/react-dropdown-menu": "^2.0.4",
@@ -49,7 +49,7 @@
     "postcss": "^8.4.21",
     "prettier": "^2.8.6",
     "prettier-plugin-tailwindcss": "^0.2.5",
-    "prisma": "^4.12.0",
+    "prisma": "^4.11.0",
     "tailwindcss": "^3.2.7",
     "typescript": "^5.0.2"
   },
